+++ src/main/java/ru/sbt/mipt/oop/Application.java
@@ -1,5 +1,7 @@
 package ru.sbt.mipt.oop;
 
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.support.AbstractApplicationContext;
 import ru.sbt.mipt.oop.Alarm.Alarm;
 import ru.sbt.mipt.oop.Event.*;
 import ru.sbt.mipt.oop.Event.SensorEvent.*;
@@ -16,48 +18,9 @@ import java.util.ArrayList;
 import java.util.Collection;
 
 public class Application{
-    public static void main(String... args) {
-        SmartHome smartHome = getSmartHome();
-        Collection<EventHandler> handlers = setEventHandlers(smartHome);
-        SensorEventsManager sensorEventsManager = new SensorEventsManager();
-        registerEventHandlers(sensorEventsManager, handlers);
+    public static void main(String[] args) {
+        AbstractApplicationContext context = new AnnotationConfigApplicationContext(MyConfiguration.class);
+        SensorEventsManager sensorEventsManager = context.getBean(SensorEventsManager.class);
         sensorEventsManager.start();
-
-        /* todo add password to alarm */
-        /* todo rewrite Adaptor */
-        /* todo do missed hw */
-        /* todo add spring impl */
-    }
-
-    private static void registerEventHandlers(SensorEventsManager sensorEventsManager, Collection<EventHandler> handlers) {
-        for(EventHandler handler: handlers) {
-            SensorEventHandlerApiAdapter apiAdapter = new SensorEventHandlerApiAdapter(handler);
-            sensorEventsManager.registerEventHandler(event_ -> apiAdapter.handle(event_));
-        }
-    }
-
-    private static SmartHome getSmartHome() {
-        SmartHomeReader reader = new SmartHomeGsonReader("smart-home-1.js");
-        return reader.Read();
-    }
-
-    private static ArrayList<NextEventGetter> setNextEventGetters() {
-        ArrayList<NextEventGetter> nextEventGetters = new ArrayList<>();
-        nextEventGetters.add(new RandomSensorNextEventGetter());
-        nextEventGetters.add(new RandomAlarmNextEventGetter());
-        return nextEventGetters;
-    }
-
-    private static Collection<EventHandler> setEventHandlers(SmartHome smartHome) {
-        Collection<EventHandler> handlers = new ArrayList<>();
-        CommandSender sender = new IOCommandSender();
-        Alarm alarm = new Alarm(smartHome, "MyStrongPassword");
-
-        handlers.add(new SecurityHandlerDecorator(new HandlerDecorator(new DoorSensorEventHandler(smartHome, sender)), alarm));
-        handlers.add(new SecurityHandlerDecorator(new HandlerDecorator(new LightSensorEventHandler(smartHome)), alarm));
-        handlers.add(new SecurityHandlerDecorator(new HandlerDecorator(new HallEventHandler(smartHome, sender)), alarm));
-
-        handlers.add(new AlarmEventHandler(alarm));
-        return handlers;
     }
 }
