+++ src/main/java/ru/sbt/mipt/oop/SensorEventHandlerApiAdapter.java
@@ -1,28 +1,8 @@
 package ru.sbt.mipt.oop;
 
-
-/*
-
-package com.coolcompany.smarthome.sample;
-
-import com.coolcompany.smarthome.events.SensorEventsManager;
-
-public class Sample {
-
-    public static void main(String[] args) {
-        SensorEventsManager sensorEventsManager = new SensorEventsManager();
-        sensorEventsManager.registerEventHandler(event -> {
-            System.out.println("Event type [" + event.getEventType() + "] from object with id=" + event.getObjectId() + "]");
-        });
-        sensorEventsManager.start();
-    }
-}
- */
-
 import ru.sbt.mipt.oop.Event.EventHandler;
 import ru.sbt.mipt.oop.Event.SensorEvent.SensorEvent;
-
-import java.util.Collection;
+import ru.sbt.mipt.oop.events.CCSensorEvent;
 
 public class SensorEventHandlerApiAdapter {
     EventHandler handler;
@@ -32,7 +12,11 @@ public class SensorEventHandlerApiAdapter {
     }
 
     void handle(Object event_){
-        SensorEvent event = ApiEvent2SensorEventAdaptor.adapt(event_);
-        handler.handleEvent(event);
+        try {
+            SensorEvent event = ApiEvent2SensorEventAdaptor.adapt(event_);
+            handler.handleEvent(event);
+        } catch (ClassCastException e){
+            System.out.println("Type <" + ((CCSensorEvent) event_).getEventType() + "> is not supported.");
+        }
     }
 }
