+++ src/main/java/ru/sbt/mipt/oop/events/SensorEventsManager.java
@@ -23,10 +23,12 @@ public class SensorEventsManager {
     }
 
     private CCSensorEvent getNextSensorEvent() {
+        //todo add Array of nextEventGetters
         // pretend like we're getting the events from physical world, but here we're going to just generate some random events
         if (Math.random() < 0.05) return null; // null means end of event stream
         String sensorEventType = eventTypes[(int) (6 * Math.random())];
         String objectId = "" + ((int) (10 * Math.random()));
+        System.out.println(sensorEventType + " " + objectId);
         return new CCSensorEvent(sensorEventType, objectId);
     }
